requirements.txt
python-telegram-bot==20.1
requests==2.28.2
m3u8==0.7.0
